+++ addOns/openapi/CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 ### Changed
-- Allow Automation Framework users to limit the number of OpenAPI messages to import (`maxMessages`). Ex: If testing authentication, access, etc.
+- Allow users to limit the number of OpenAPI messages to import (`maxMessages`) via the Automation Framework, Import dialogue, and API. Ex: If testing authentication, access, etc.
 
 ## [57] - 2026-07-06
 ### Changed